Renamin node_filter to nodes_filter.
